Add support for KSP-AVC.
This file lets players use either the KSP-AVC utility by @tyrope or the KSP-AVC plugin by @CYBUTEK to check for new releases of RemoteTech. This file will have to be updated with the version number of the latest release (WARNING: do not change the version number on `master` until the release is ready).

Closes Cilph/RemoteTech2#237. See also RemoteTechnologiesGroup/RemoteTech#33 and Cilph/RemoteTech2/#217.