Added Rover Computer to the Flight Computer
Largely based on the implementation for the old RT2 repo.
Rover Computer provides 3 modes:
-Drive, for driving a set distance or turning a set number of degrees
-Heading. For keaping a heading for a set distance
-Coord. For driving autonomously to a set coordinate.
Coordinates can be typed in manually (pressing enter in any field will
send the command) or the player can <modifierKey>+left click on the
ground in either flight or map view to plot in a representation of the
clicked spot in coordinates. <modifierKey>+left click does the same, it
just sends the command automatically.