Fixed Flight Computer for vessels without Reaction Wheels.
This fix consists basically only of a gimbal vectoring calculation update. The old code seemed to be from an earlier version of KSP.
The bug manifested itself in a craft that would not cahnge orientation whatsoever when the Flight computer was enabled. It only started to jerk around and kept facing the orientation it was facing in. This was independent of the mode of the Flight Computer. The bug can be easily reproduced by adding the following .cfg file and using the Flight Computer on the Stayputnik probe core:

@PART[probeCoreSphere]
{
  !MODULE[ModuleReactionWheel]
  {}
}

I encountered this bug while playing with the RealismOverhaul mod, which removes from the reaction wheels from essentially all early tech probes.

I also fixed a divide by zero bug that crashed KSP, when using the above setup.

No mod other than RT2 and the custom .cfg file mentioned above are required to reproduce the bug.