Updated to KSP-AVC Plugin 1.1.3 or later.
KSP-AVC Utility is now defunct, and doesn't need to be supported.

So long as we use release tags that include the version number in them, we don't need to worry about race conditions between updating the online .version file and actually publishing the release; the version inferred from the latest release tag will override the version in the .version file.