Added save file for testing RemoteTech.
A standardized file should help with regression testing. I suggest that other developers update this file as needed when they find new situations to test. The current file only has vessels within Kerbin's sphere of influence, but this is enough to test dish targeting, communication between different spheres of influence, and the behavior of Active Vessel targeting.

This directory will not appear in the Travis build, which only packages `GameData/`.