Moved `GetPositionFromGuid()` to `NetworkManager`.
Just realized that RangeModelExtensions is meant to have, yes, extension methods. `GetPositionFromGuid()` doesn't qualify, but it serves a similar purpose to a few other `NetworkManager` methods.

Throwing an exception on invalid input is not standard for RemoteTech methods, but since Vector3d is not nullable I can't think of a good way to handle the error internally.