Small fixes
I changed the saving method for each active and queued command. We
always passed the complete `flighcomputer`-Node to each command for
saving. So this can cause problems by saving a command-data to his own
node.
The `ExternalApiCommand` will now save the passed externalData
ConfigNode from the API call to the persistent correctly.