CancelCommand refactoring and changes
I've create a uniqued number for each `AbstractCommand`. The
`CancelCommand` will now cancel the command with its unique number and
no longer with the position. I also changed the loading method. We'll no
longer search the command to cancel within the loading method, because
this can cause, that we can't find the command to cancel when the
CancelCommand is loaded first.