EC format changed when lower than 1.00/s
For RSS/RO/RP-0 players the eletric consumption of the long antenna is
displayed as 0.00/s because we've a formated string with F2 and for this
antenna the consumption is 0.0015/s. We'll now tranform all consumptions
lower than 1.00/s into minutes. I've also changed the label `Energy
req.` into a more stock like label `ElectricCharge`. Also `s` to sec and
min for minutes