MissionControl Update pack
##### General
- The mission control dot on the map view will now be invisible on a
defined distance to the camera. See `DistanceToHideGroundStations` on
the settings file. This feature can be toggled by the value
`HideGroundStationsOnDistance` in the option window.
- I've added a mouse over info box for each mission control dot on the
map view. This feature can be toggled by the value
`ShowMouseOverInfoGroundStations` in the option window.

##### Modders
- Ground station antennas can now use the tech level of the tracking
station. I've added three new values to the antenna node for a station.
Use `UpgradeableOmni`, `UpgradeableDish` and `UpgradeableCosAngle` with
a semicolon seperated list of ranges. Example: 4E+06;3.0E+07;7.5E+07

##### Contributors
- I've added three trigger events for the RTSettings.
`OnSettingsChanged` will be fired when the option window is closed.
`OnSettingsLoaded` will be fired when the settings is fully loaded, and
`OnSettingsSaved` will be fired by saving the rtsettings file.