Update to use triggered flag on science data
This checks for the triggered flag before creating the commStream. If the commStream is left null then the process skips the part about pushing data to the R&D center. Instead, after transmission is completed the triggered data transmission event is fired.

An additional bool is added for aborted transmissions, so that the triggered data event can send the correct arguments.